+++ ps_funcs/py/__init__.py
@@ -15,7 +15,7 @@ from . import shfuncs
 from . import path
 from . import cli
 
-#from . import files
+from . import files
 from . import ps_csv
 from . import structs
 from . import strings
@@ -120,34 +120,14 @@ def file_get_contents (filename, binary=False, forceZip=False, return_lst_bool:b
 	Function mimics PHP file_get_contents minimal behavior - reads file specified by *filename* arg and return its content.
 	However it is not fully compatible substitution cause it doesn't supports any psrameters beside *filename*
 	"""
-	try:
-		res=None
-		with open (filename, "rb" if binary else "rt") as inf:
-			if return_lst_bool:
-				res=strip_list_items(inf.readlines())
-			else:
-				res=inf.read()
-		return res
-	except FileNotFoundError:
-		raise PSException ("File", filename, "not found", fatal=True)
-	return None
+	return files.file_get_contents (filename, binary, forceZip, return_lst_bool)
 	
-def file_put_contents (filename, data, overwrite=False, return_none_on_exists=False)->int|None:
+def file_put_contents (filename_fnstr, data, overwrite=False, return_none_on_exists=False)->int|None:
 	"""
 	Function mimics PHP file_put_contents minimal behavior - reads file specified by *filename* arg and return its content.
 	However it is not fully compatible substitution cause it doesn't supports any psrameters beside *filename* and *data*
 	"""
-	try:
-		outf=open (filename, 'w' if overwrite else 'x')
-		written_bytes_int=outf.write(data)
-		outf.close()
-		return written_bytes_int
-	except Exception as e:
-		if isinstance(e, FileExistsError):
-			return None
-		else:
-			return 0
-		
+	return files.file_put_contents (filename_fnstr, data, overerite, return_none_on_exists)
 
 # [@ps_section]  structures helpers