+++ ps_funcs/py/structs.py
@@ -1,0 +1,21 @@
+"""
+Module with functions for working with
+various structurrs like dicts, lists, tuples etc.
+"""
+
+
+# @ps section list functions
+def list_get (lst_list, index_int: int, default=None):
+	"""
+	Safe get value of list by index returning default value if index out of range
+	"""
+	try:
+		return lst_list[index_int]
+	except:
+		return default
+
+def strip_list_items (lst:list)->list:
+	"""
+	Apply `strip` method to every element in list
+	"""
+	return [li.strip() for li in lst]